requirements.txt
gradio
fastapi
python-dotenv
langchain
langchain_groq
uvicorn
pydantic